+++ mobile/lib/blocs/dashboard/dashboard_bloc.dart
@@ -1,14 +1,13 @@
 import 'dart:async';
 
 import 'package:app/models/models.dart';
-import 'package:app/utils/extensions.dart';
+import 'package:app/services/services.dart';
+import 'package:app/utils/utils.dart';
 import 'package:equatable/equatable.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import 'package:hive/hive.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 
-import '../../services/hive_service.dart';
-import '../../utils/dashboard.dart';
 
 part 'dashboard_event.dart';
 part 'dashboard_state.dart';
