+++ mobile/lib/screens/auth/email_auth_widget.dart
@@ -12,7 +12,9 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../on_boarding/on_boarding_widgets.dart';
 import '../on_boarding/profile_setup_screen.dart';
+import 'auth_verification.dart';
 import 'auth_widgets.dart';
 
 class EmailAuthWidget extends StatefulWidget {
