+++ mobile/lib/screens/auth/phone_auth_widget.dart
@@ -11,7 +11,9 @@ import 'package:auto_size_text/auto_size_text.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
+import '../on_boarding/on_boarding_widgets.dart';
 import '../on_boarding/profile_setup_screen.dart';
+import 'auth_verification.dart';
 import 'auth_widgets.dart';
 import 'email_auth_widget.dart';
