+++ mobile/lib/services/firebase_service.dart
@@ -1,7 +1,9 @@
 import 'dart:io';
 
+import 'package:app/blocs/blocs.dart';
 import 'package:app/constants/constants.dart';
 import 'package:app/models/models.dart';
+import 'package:app/services/services.dart';
 import 'package:app/utils/utils.dart';
 import 'package:app/widgets/widgets.dart';
 import 'package:cloud_firestore/cloud_firestore.dart';
@@ -637,7 +639,7 @@ class CustomAuth {
   }
 
   static Future<void> sendPhoneAuthCode(
-    phoneNumber,
+    String phoneNumber,
     BuildContext buildContext,
   ) async {
     try {
@@ -685,7 +687,7 @@ class CustomAuth {
   }
 
   static Future<void> sendEmailAuthCode(
-    emailAddress,
+    String emailAddress,
     BuildContext buildContext,
   ) async {
     try {
