+++ mobile/lib/widgets/custom_widgets.dart
@@ -382,10 +382,12 @@ class AppSafeArea extends StatelessWidget {
     super.key,
     required this.widget,
     this.verticalPadding,
+    this.horizontalPadding,
     this.backgroundColor,
   });
   final Widget widget;
   final double? verticalPadding;
+  final double? horizontalPadding;
   final Color? backgroundColor;
 
   @override
@@ -393,7 +395,10 @@ class AppSafeArea extends StatelessWidget {
     return Container(
       color: backgroundColor ?? CustomColors.appBodyColor,
       child: SafeArea(
-        minimum: EdgeInsets.symmetric(vertical: verticalPadding ?? 0),
+        minimum: EdgeInsets.symmetric(
+          vertical: verticalPadding ?? 0,
+          horizontal: horizontalPadding ?? 0,
+        ),
         child: widget,
       ),
     );
